+++ ntex-util/src/lib.rs
@@ -7,7 +7,6 @@
 )]
 
 pub mod channel;
-pub mod error;
 pub mod future;
 pub mod services;
 pub mod task;
@@ -19,3 +18,8 @@ pub use ntex_rt::spawn;
 pub type HashMap<K, V> = std::collections::HashMap<K, V, foldhash::fast::RandomState>;
 pub type HashSet<V> = std::collections::HashSet<V, foldhash::fast::RandomState>;
 pub type HashRandomState = foldhash::fast::RandomState;
+
+#[deprecated(since = "3.5.0", note = "Use ntex-error 1.2")]
+pub mod error {
+    pub use ntex_error::{ErrorMessage, ErrorMessageChained, fmt_err, fmt_err_string};
+}