+++ ntex-error/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [1.2.0] - 2026-03-14
+
+* Refactor ErrorDiagnostic trait
+
+* Create ErrorInformation from Error reference
+
 ## [1.1.0] - 2026-03-09
 
 * Introduce ErrorInformation type